Categorize publications on homepage using the widget
I don't separate working papers in a new content category anymore. I use
the home page widgets to select the publication type to show. This allows
me to also have a policy brief category. This also fixes the publication
search page as it was broken before with working papers as a different
publication category as other publications.